+++ src/core/models/setup.ts
@@ -1,5 +1,6 @@
 import { ICuiEventBusQueueSetup } from "../bus/interfaces";
 import { CuiLogLevel, CuiInteractionsType } from "../utils/types";
+import { ICuiDevelopmentTool } from "./interfaces";
 
 interface CuiSetupCommon {
     prefix?: string;
@@ -61,6 +62,7 @@ export class CuiSetupInit implements CuiSetupCommon {
     scrollThreshold?: number;
     resizeThreshold?: number;
     busSetup?: ICuiEventBusQueueSetup[];
+    development: ICuiDevelopmentTool;
     root: HTMLElement;
 
     constructor() {